+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Basic implementations of the components `Backdrop`, `BaseModal` and `Modal`.
 
 ## [0.2.1] - 2019-11-19
 
@@ -16,6 +18,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test example for VTEX IO app.
 
 ## Changed
-- Removed `injectIntl` from example
+- Removed `injectIntl` from example.
 
 - **Component** Create the VTEX Store Component _IO Base App_